+++ Ximea/dialogXimea.h
@@ -56,8 +56,8 @@ class DialogXimea : public ito::AbstractAddInConfigDialog
 		bool m_inEditing;
 		void enableDialog(bool enabled);
 
-        inline double msecToSec(int musec) { return (double)musec * 1.0e-3; }
-		inline int secToMsec(double sec) { return (int)(sec * 1.0e3); }
+        inline double msecToSec(double musec) { return (double)musec * 1.0e-3; }
+		inline double secToMsec(double sec) { return (double)(sec * 1.0e3); }
 
     public slots: 
 		void parametersChanged(QMap<QString, ito::Param> params);
